core/requirements.txt
django = 4.1.1
django_filter = 22.1